+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Use new Share UI for bookmarks sharing [#2656](https://github.com/Automattic/pocket-casts-ios/pull/2656)
 - Remove the option to share bookmarks from users file [#2661](https://github.com/Automattic/pocket-casts-ios/pull/2661)
 - Add direct access to sleep timer settings from player [#2698](https://github.com/Automattic/pocket-casts-ios/pull/2698)
+- Add option to remove selected episodes from your listening history [#2702](https://github.com/Automattic/pocket-casts-ios/pull/2702)
 
 7.80
 -----