+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 7.82
 -----
-
+- Add App Clips [#2549](https://github.com/Automattic/pocket-casts-ios/issues/2549)
 
 7.81
 -----
 - Use new Share UI for bookmarks sharing [#2656](https://github.com/Automattic/pocket-casts-ios/pull/2656)
 - Remove the option to share bookmarks from users file [#2661](https://github.com/Automattic/pocket-casts-ios/pull/2661)
-- Add App Clips [#2549](https://github.com/Automattic/pocket-casts-ios/issues/2549)
 
 7.80
 -----