+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Remove the option to share bookmarks from users file [#2661](https://github.com/Automattic/pocket-casts-ios/pull/2661)
 - Add direct access to sleep timer settings from player [#2698](https://github.com/Automattic/pocket-casts-ios/pull/2698)
 - Add option to remove selected episodes from your listening history [#2702](https://github.com/Automattic/pocket-casts-ios/pull/2702)
+- Fixed slow loading of long bookmarks lists [#2699](https://github.com/Automattic/pocket-casts-ios/pull/2699)
 
 7.80
 -----