+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 7.82
 -----
-
+- Add App Clips [#2549](https://github.com/Automattic/pocket-casts-ios/issues/2549)
 
 7.81
 -----